+++ Assets/CommonCore/Core/CoreParams.cs
@@ -37,6 +37,7 @@ namespace CommonCore
 
         public static bool SetSafeResolutionOnExit { get; private set; } = true;
         public static Vector2Int SafeResolution { get; private set; } = new Vector2Int(1280, 720);
+        public static bool SetNativeResolutionOnFirstRun { get; private set; } = true;
 
         public static UIThemePolicy UIThemeMode { get; private set; } = UIThemePolicy.ExplicitOnly;
         public static string DefaultUITheme { get; private set; } = "ThresholdTheme";